dotbot/examples/qrkey_demo: relocate from dotbot/qrkey
Depends on the upstream qrkey package; not vendored.

AI-assisted: Claude Opus 4.7
