+++ dotbot/examples/qrkey_demo/__main__.py
@@ -1,0 +1,9 @@
+# SPDX-FileCopyrightText: 2026-present Inria
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""Entry point for `python -m dotbot.examples.qrkey_demo`."""
+
+from dotbot.examples.qrkey_demo.cli import main
+
+if __name__ == "__main__":
+    main()  # pylint: disable=no-value-for-parameter
