+++ dotbot/tests/test_qrkey.py
@@ -5,8 +5,8 @@ from unittest.mock import AsyncMock, MagicMock
 import pytest
 from websockets import exceptions as websockets_exceptions
 
+from dotbot.examples.qrkey_demo import QrKeyClient, QrKeyClientSettings
 from dotbot.logger import setup_logging
-from dotbot.qrkey import QrKeyClient, QrKeyClientSettings
 
 
 class WebsocketMock:
@@ -33,7 +33,8 @@ def client(monkeypatch):
     qrkey_controller_mock = MagicMock()
     qrkey_controller_mock.start.side_effect = qrkey_controller_start_mock
     monkeypatch.setattr(
-        "dotbot.qrkey.QrkeyController", lambda *args, **kwargs: qrkey_controller_mock
+        "dotbot.examples.qrkey_demo.client.QrkeyController",
+        lambda *args, **kwargs: qrkey_controller_mock,
     )
     websocket_mock = WebsocketMock()
 
@@ -42,9 +43,12 @@ def client(monkeypatch):
         return json.dumps({"cmd": 2, "data": {"address": "test_bot"}})
 
     websocket_mock.recv.side_effect = recv_side_effect
-    monkeypatch.setattr("dotbot.qrkey.connect", lambda *args, **kwargs: websocket_mock)
+    monkeypatch.setattr(
+        "dotbot.examples.qrkey_demo.client.connect",
+        lambda *args, **kwargs: websocket_mock,
+    )
     rest_client = MagicMock()
-    monkeypatch.setattr("dotbot.qrkey.RestClient", rest_client)
+    monkeypatch.setattr("dotbot.examples.qrkey_demo.client.RestClient", rest_client)
 
     settings = QrKeyClientSettings(
         http_port=8001,
