+++ pyproject.toml
@@ -71,7 +71,9 @@ dotbot-controller = "dotbot.controller_app:main"
 dotbot-edge-gateway = "dotbot.edge_gateway_app:main"
 dotbot-keyboard = "dotbot.keyboard:main"
 dotbot-joystick = "dotbot.joystick:main"
-dotbot-qrkey = "dotbot.qrkey_app:main"
+# dotbot-qrkey console script removed — the demo is now an example,
+# run via `python -m dotbot.examples.qrkey_demo`. See
+# dotbot/examples/qrkey_demo/README.md.
 
 [tool.ruff]
 lint.select = ["E", "F"]
