+++ tox.ini
@@ -51,7 +51,6 @@ allowlist_externals=
     /usr/bin/bash
 commands=
     bash -exc "dotbot-controller --help"
-    bash -exc "dotbot-qrkey --help"
 
 [testenv:web]
 allowlist_externals=
